moved event creation to observers so that caching is not required for development mode
git-svn-id: http://svn.techno-weenie.net/projects/mephisto/trunk@1427 567b1171-46fb-0310-a4c9-b4bef9110e78

+++ app/cachers/template_sweeper.rb
@@ -8,6 +8,7 @@ class TemplateSweeper < ActionController::Caching::Sweeper
   end
 
   def after_save(record)
+    return if controller.nil?
     controller.class.benchmark "Expired all referenced pages" do
       CachedPage.find(:all).each { |p| controller.class.expire_page(p.url) }
       CachedPage.delete_all
